Region-Grouped 3D Mesh
Alrighty, 3D now works in full colour and with the voxels being grouped based on their given region.

The code is currently very poorly optimised all over the shop but in regards to the 3D Mesh script, it is less optimised than an early access money grab.

So, with that said I recommend only using 3D currently to render out small terrain maps